# refactor recomputation to work with tags