Another step of #1560 

This refactors the recomputation (activation checkpointing):
- recomputation works with tags,
- automatically tag intermediates from autograd on decomposition as to be recomputed, (this matches the behaviour of rematerialize_forward_backward, I think),
- I needed to disable rematerialize_forward_backward because it ran into "infinite capacity". However, I think this is not needed after this PR (cc @IvanYashchuk),
- move the uniform -> get_and_update_random_state + uniform_philox transform to before the autograd
- guard against recomputing random ops,

This is expected to be memory/compute neutral (I'll report numbers in a bit). It does not yet do the checkpointing frontend for the jit (including using memory commparable to eager checkpointing), that will be a separate PR.

@riccardofelluga could you take a look? (should be no surprises relative to #1560)