@t-vi 2025-01-07
So there is quite a regression in memory use:

```
python thunder/benchmarks/benchmark_litgpt.py --model_name stablecode-completion-alpha-3b --compile thunder --checkpoint_activations True --low_precision_mode none --micro_batch_size 1 --n_layer 4 --max_iters 3 --warmup_iters 2  --dump_thunder_traces True 
```
gives

- Main: Average iter time: 1279.73 ms, Memory used: 13.27 GB
- This PR: Average iter time: 1139.99 ms Memory used: 16.23 GB

So 10% faster but 3GB of memory. 
This seems to be because we don't recompute everything. I'll look into it.