@t-vi 2025-01-07
So I'm not entirely happy with disabling the `rematerialize_forward_and_backward`, but I wonder if the need arises from proxies that are re-computed and thus appear twice in the `joint_extrace`. (Hello name collision.)
I'll try that either here or in a follow-up.