@t-vi 2025-01-08
Unfortunately, disabling the forward-backward-rematerialization adversely affects memory for Qwen:

```
python thunder/benchmarks/benchmark_litgpt.py --model_name Qwen2.5-7B --compile thunder --checkpoint_activations True --low_precision_mode none --micro_batch_size 1 --n_layer 4 --max_iters 3 --warmup_iters 2 --block_size 4096 
```
- Main: Average iter time: 788.88 ms, Memory used: 18.76 GB
- This PR: Average iter time: 686.02 ms, Memory used: 20.77 GB
- PR #1616 : Average iter time: 779.80 ms, Memory used: 18.34 GB
- Eager instead of thunder: Average iter time: 803.95 ms Memory used: 17.55 GB