@IvanYashchuk 2025-01-08
> automatically tag intermediates from autograd on decomposition as to be recomputed, (this matches the behaviour of rematerialize_forward_backward, I think),

No, it doesn't match the behavior of `rematerialize_forward_backward`. What to recompute in fusion regions is decided by a min-cut-based algorithm. This PR introduced a regression in peak used memory (checked for Llama 2 7B): https://github.com/Lightning-AI/lightning-thunder/issues/1621. 

> I needed to disable rematerialize_forward_backward because it ran into "infinite capacity".

The rematerialization code has assumptions on input traces to be functioning, these assumptions were violated resulting in the "infinite capacity" error. @riccardofelluga was hitting the same problem when working on the recomputation. The problems are supposed to be fixed with https://github.com/Lightning-AI/lightning-thunder/pull/1367. Riccardo, what's the current status of 1367?