@riccardofelluga 2025-01-09
#1367 was parked due to change of priorities, we could tho bring that back here adapting to the new saved-for-backward logic of #1615. Tho it can also be that the infinite capacity error is reached by a different cause